Course format tests are self contained (no need to split into cleaner, uncleaner and clean_unclean).
